Phoenix v1.4.15 mix phx.new.ecto

Creates a new Ecto project within an umbrella project.

This task is intended to create a bare Ecto project without web integration, which serves as a core application of your domain for web applications and your greater umbrella platform to integrate with.

It expects the name of the project as an argument.

$ cd my_umbrella/apps
$ mix phx.new.ecto APP [--module MODULE] [--app APP]

A project at the given APP directory will be created. The application name and module name will be retrieved from the application name, unless --module or --app is given.

Options

--app - the name of the OTP application

--module - the name of the base module in the generated skeleton

--database - specify the database adapter for Ecto. One of:

postgres (https://github.com/elixir-ecto/postgrex)

mysql (https://github.com/elixir-ecto/myxql)

Please check the driver docs, between parentheses, for more information and requirements. Defaults to "postgres".

--binary-id - use binary_id as primary key type in Ecto schemas

Examples

mix phx.new.ecto hello_ecto

Is equivalent to:

mix phx.new.ecto hello_ecto --module HelloEcto

Functions

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.